+++ src/main/java/io/appform/dropwizard/actors/actor/HaModeVisitor.java
@@ -1,0 +1,8 @@
+package io.appform.dropwizard.actors.actor;
+
+public interface HaModeVisitor<T> {
+
+    T visitAll();
+
+    T visitExactly();
+}