+++ src/test/java/io/appform/dropwizard/actors/connectivity/actor/NamespacedQueuesTest.java
@@ -1,5 +1,9 @@
 package io.appform.dropwizard.actors.connectivity.actor;
 
+import static io.appform.dropwizard.actors.utils.RMQContainer.RABBITMQ_MANAGEMENT_PORT;
+import static io.appform.dropwizard.actors.utils.RMQContainer.RABBITMQ_PASSWORD;
+import static io.appform.dropwizard.actors.utils.RMQContainer.RABBITMQ_USERNAME;
+
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.health.HealthCheckRegistry;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -17,63 +21,66 @@ import io.appform.dropwizard.actors.metrics.RMQMetricObserver;
 import io.appform.dropwizard.actors.retry.RetryStrategyFactory;
 import io.appform.dropwizard.actors.utils.ActorType;
 import io.appform.dropwizard.actors.utils.AsyncOperationHelper;
+import io.appform.dropwizard.actors.utils.RMQContainer;
 import io.appform.dropwizard.actors.utils.SidelineTestActor;
 import io.appform.dropwizard.actors.utils.TestMessage;
-import io.appform.testcontainers.rabbitmq.RabbitMQStatusCheck;
-import io.appform.testcontainers.rabbitmq.config.RabbitMQContainerConfiguration;
 import io.dropwizard.Configuration;
 import io.dropwizard.jackson.Jackson;
 import io.dropwizard.setup.Environment;
 import io.dropwizard.testing.junit5.DropwizardAppExtension;
-import lombok.SneakyThrows;
-import lombok.extern.slf4j.Slf4j;
-import lombok.val;
-import okhttp3.*;
-import org.junit.*;
-import org.testcontainers.containers.GenericContainer;
-
-import javax.validation.Validation;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.Executors;
+import javax.validation.Validation;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import lombok.val;
+import okhttp3.Credentials;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.RequestBody;
+import okhttp3.Response;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.RabbitMQContainer;
 
 
 @Slf4j
 public class NamespacedQueuesTest {
 
-    private static final int RABBITMQ_MANAGEMENT_PORT = 15672;
-    private static final String RABBITMQ_DOCKER_IMAGE = "rabbitmq:3.8.34-management";
-    private static final String RABBITMQ_USERNAME = "guest";
-    private static final String RABBITMQ_PASSWORD = "guest";
     private static final String NAMESPACE_VALUE = "namespace1";
 
     public static final DropwizardAppExtension<RabbitMQBundleTestAppConfiguration> app =
             new DropwizardAppExtension<>(RabbitMQBundleTestApplication.class);
     private final MetricRegistry metricRegistry = new MetricRegistry();
 
-    @BeforeClass
+    @BeforeAll
     @SneakyThrows
     public static void setupClass() {
         app.before();
     }
 
-    @AfterClass
+    @AfterAll
     @SneakyThrows
     public static void cleanupClass() {
         app.after();
     }
 
-    @Before
-    public void setup() {
-        System.setProperty(NamingUtils.NAMESPACE_PROPERTY_NAME, NAMESPACE_VALUE);
+    private static RabbitMQContainer rabbitMQContainer() {
+        return RMQContainer.startContainer();
     }
 
-    @After
-    public void cleanup() {
-        System.setProperty(NamingUtils.NAMESPACE_PROPERTY_NAME, "");
+    @BeforeEach
+    public void setup() {
+        System.setProperty(NamingUtils.NAMESPACE_PROPERTY_NAME, NAMESPACE_VALUE);
     }
     private RMQConfig config;
 
@@ -86,59 +93,69 @@ public class NamespacedQueuesTest {
      */
     @Test
     public void testQueuesAreNamespacedWhenFeatureEnvIsSet() throws Exception {
-        GenericContainer rabbitMQContainer = rabbitMQContainer();
+        String queueName = "publisher-0";
+        RabbitMQContainer rabbitMQContainer = rabbitMQContainer();
         val mappedManagementPort = rabbitMQContainer.getMappedPort(RABBITMQ_MANAGEMENT_PORT);
         config = getRMQConfig(rabbitMQContainer);
 
-        RMQConnection connection = new RMQConnection("test-conn", config,
+        RMQConnection connection = new RMQConnection("test-conn-0", config,
                 Executors.newSingleThreadExecutor(), app.getEnvironment(), TtlConfig.builder().build(), new RMQMetricObserver(config, metricRegistry));
         connection.start();
 
         ActorConfig actorConfig = new ActorConfig();
-        actorConfig.setExchange("test-exchange-1");
-        UnmanagedPublisher publisher = new UnmanagedPublisher<>(
-                "publisher-1", actorConfig, connection, null);
+        actorConfig.setExchange("test-exchange-0");
+        UnmanagedPublisher publisher = new UnmanagedPublisher<>(queueName, actorConfig, connection, null);
         publisher.start();
 
         ObjectMapper objectMapper = new ObjectMapper();
         Response response = sendRequest("/api/queues", mappedManagementPort);
+        String queueNameWithNamespace = String.format("%s.%s.%s", NAMESPACE_VALUE, actorConfig.getPrefix(), queueName);
         if (response != null) {
             JsonNode jsonNode = objectMapper.readTree(response.body().string());
             if (jsonNode.isArray()) {
                 for (JsonNode json : jsonNode) {
-                    String queueName = json.get("name").asText();
-                    Assert.assertTrue(queueName.contains(NAMESPACE_VALUE));
+                    String actualQueueName = json.get("name").asText();
+                    if (actualQueueName.endsWith(queueName)) {
+                        Assertions.assertEquals(queueNameWithNamespace, actualQueueName);
+                    }
                 }
             }
             response.close();
         }
     }
 
+    @AfterEach
+    public void cleanup() {
+        System.clearProperty(NamingUtils.NAMESPACE_PROPERTY_NAME);
+    }
+
     @Test
     public void testQueuesAreNotNamespacedWhenFeatureEnvNotSet() throws Exception {
-        System.setProperty(NamingUtils.NAMESPACE_PROPERTY_NAME, "");
-        GenericContainer rabbitMQContainer = rabbitMQContainer();
+        String queueName = "publisher-1";
+        System.clearProperty(NamingUtils.NAMESPACE_PROPERTY_NAME);
+        RabbitMQContainer rabbitMQContainer = rabbitMQContainer();
         val mappedManagementPort = rabbitMQContainer.getMappedPort(RABBITMQ_MANAGEMENT_PORT);
         config = getRMQConfig(rabbitMQContainer);
 
-        RMQConnection connection = new RMQConnection("test-conn", config,
+        RMQConnection connection = new RMQConnection("test-conn-1", config,
                 Executors.newSingleThreadExecutor(), app.getEnvironment(), TtlConfig.builder().build(), new RMQMetricObserver(config, metricRegistry));
         connection.start();
 
         ActorConfig actorConfig = new ActorConfig();
         actorConfig.setExchange("test-exchange-1");
-        UnmanagedPublisher publisher = new UnmanagedPublisher<>(
-                "publisher-1", actorConfig, connection, null);
+        UnmanagedPublisher publisher = new UnmanagedPublisher<>(queueName, actorConfig, connection, null);
         publisher.start();
 
         ObjectMapper objectMapper = new ObjectMapper();
         Response response = sendRequest("/api/queues", mappedManagementPort);
+        String queueNameWithNamespace = String.format("%s.%s.%s", NAMESPACE_VALUE, actorConfig.getPrefix(), queueName);
         if (response != null) {
             JsonNode jsonNode = objectMapper.readTree(response.body().string());
             if (jsonNode.isArray()) {
                 for (JsonNode json : jsonNode) {
-                    String queueName = json.get("name").asText();
-                    Assert.assertFalse(queueName.contains(NAMESPACE_VALUE));
+                    String actualQueueName = json.get("name")
+                            .asText();
+                    Assertions.assertNotEquals(actualQueueName, queueNameWithNamespace);
                 }
             }
             response.close();
@@ -147,41 +164,93 @@ public class NamespacedQueuesTest {
 
     @Test
     public void testQueuesAreRemovedAfterTtl() throws Exception {
-        GenericContainer rabbitMQContainer = rabbitMQContainer();
+        String queueName = "publisher-2";
+        RabbitMQContainer rabbitMQContainer = rabbitMQContainer();
         val mappedManagementPort = rabbitMQContainer.getMappedPort(RABBITMQ_MANAGEMENT_PORT);
         config = getRMQConfig(rabbitMQContainer);
 
         TtlConfig ttlConfig = TtlConfig.builder()
                 .ttlEnabled(true)
-                .ttl(Duration.ofSeconds(5))
+                .ttl(Duration.ofSeconds(2))
                 .build();
-        RMQConnection connection = new RMQConnection("test-conn", config,
+        RMQConnection connection = new RMQConnection("test-conn-2", config,
                 Executors.newSingleThreadExecutor(), app.getEnvironment(), ttlConfig, new RMQMetricObserver(config, metricRegistry));
         connection.start();
 
         ActorConfig actorConfig = new ActorConfig();
-        actorConfig.setExchange("test-exchange-1");
-        UnmanagedPublisher publisher = new UnmanagedPublisher<>(
-                "publisher-1", actorConfig, connection, null);
+        actorConfig.setExchange("test-exchange-2");
+        UnmanagedPublisher publisher = new UnmanagedPublisher<>(queueName, actorConfig, connection, null);
         publisher.start();
 
-        Thread.sleep(10000);
+        Thread.sleep(4_000);
         Response response = sendRequest("/api/queues", mappedManagementPort);
         if (response != null) {
             ObjectMapper objectMapper = new ObjectMapper();
             JsonNode jsonNode = objectMapper.readTree(response.body().string());
-            Assert.assertTrue(jsonNode.size() == 0);
+            if (jsonNode.isArray()) {
+                for (JsonNode json : jsonNode) {
+                    String actualQueueName = json.get("name")
+                            .asText();
+                    Assertions.assertFalse(actualQueueName.contains(queueName));
+                }
+            }
             response.close();
         }
     }
 
+    private RMQFetchMessages getBody() {
+        return new RMQFetchMessages();
+    }
+
+    private Response sendRequest(String endpoint,
+                                 int mappedManagementPort) {
+        OkHttpClient client = new OkHttpClient();
+        String credential = Credentials.basic(config.getUserName(), config.getPassword());
+        Request request = new Request.Builder().url("http://" + config.getBrokers()
+                        .get(0)
+                        .getHost() + ":" + mappedManagementPort + endpoint)
+                .header("Accept", "application/json")
+                .header("Authorization", credential)
+                .build();
+
+        try {
+            return client.newCall(request)
+                    .execute();
+        } catch (IOException e) {
+            log.error("Error while making API call to RabbitMQ", e);
+        }
+        return null;
+    }
+
+    private Response sendPostRequest(String endpoint,
+                                     String body,
+                                     int mappedManagementPort) {
+        OkHttpClient client = new OkHttpClient();
+        String credential = Credentials.basic(config.getUserName(), config.getPassword());
+        Request request = new Request.Builder().url("http://" + config.getBrokers()
+                        .get(0)
+                        .getHost() + ":" + mappedManagementPort + endpoint)
+                .header("Accept", "application/json")
+                .header("Authorization", credential)
+                .post(RequestBody.create(body.getBytes(StandardCharsets.UTF_8)))
+                .build();
+
+        try {
+            return client.newCall(request)
+                    .execute();
+        } catch (IOException e) {
+            log.error("Error while making API call to RabbitMQ", e);
+        }
+        return null;
+    }
+
     @Test
     public void testQueuesSidelineForFailedMessages() throws Exception {
-        GenericContainer rabbitMQContainer = rabbitMQContainer();
+        RabbitMQContainer rabbitMQContainer = rabbitMQContainer();
         val mappedManagementPort = rabbitMQContainer.getMappedPort(RABBITMQ_MANAGEMENT_PORT);
         config = getRMQConfig(rabbitMQContainer);
 
-        RMQConnection connection = new RMQConnection("test-conn", config,
+        RMQConnection connection = new RMQConnection("test-conn-3", config,
                 Executors.newSingleThreadExecutor(), app.getEnvironment(), null, new RMQMetricObserver(config, metricRegistry));
         connection.start();
 
@@ -211,75 +280,17 @@ public class NamespacedQueuesTest {
         String endpoint = "/api/queues/%2F/" + sidelineQueue + "/get";
 
         Response response = sendPostRequest(endpoint, objectMapper.writeValueAsString(getBody()), mappedManagementPort);
-        Assert.assertNotNull(response);
+        Assertions.assertNotNull(response);
 
         JsonNode jsonNode = objectMapper.readTree(response.body().string());
-        Assert.assertEquals( 1, jsonNode.size());
+        Assertions.assertEquals(1, jsonNode.size());
         JsonNode messageResponse = jsonNode.get(0);
-        Assert.assertEquals("test.exchange_SIDELINE", messageResponse.get("exchange").asText());
+        Assertions.assertEquals("test.exchange_SIDELINE", messageResponse.get("exchange")
+                .asText());
         TestMessage actualMessage = objectMapper.readValue(messageResponse.get("payload").asText(), TestMessage.class);
-        Assert.assertEquals(ActorType.ALWAYS_FAIL_ACTOR, actualMessage.getActorType());
-        Assert.assertEquals("test_message", actualMessage.getName());
+        Assertions.assertEquals(ActorType.ALWAYS_FAIL_ACTOR, actualMessage.getActorType());
+        Assertions.assertEquals("test_message", actualMessage.getName());
         response.close();
-    }
-
-    private RMQFetchMessages getBody() {
-        return new RMQFetchMessages();
-    }
-
-    private Response sendRequest(String endpoint, int mappedManagementPort) {
-        OkHttpClient client = new OkHttpClient();
-        String credential = Credentials.basic(config.getUserName(), config.getPassword());
-        Request request = new Request.Builder()
-                .url("http://" + config.getBrokers().get(0).getHost() + ":" + mappedManagementPort + endpoint)
-                .header("Accept", "application/json")
-                .header("Authorization", credential)
-                .build();
-
-        try {
-            return client.newCall(request).execute();
-        } catch (IOException e) {
-            log.error("Error while making API call to RabbitMQ", e);
-        }
-        return null;
-    }
-
-    private Response sendPostRequest(String endpoint, String body, int mappedManagementPort) {
-        OkHttpClient client = new OkHttpClient();
-        String credential = Credentials.basic(config.getUserName(), config.getPassword());
-        Request request = new Request.Builder()
-                .url("http://" + config.getBrokers().get(0).getHost() + ":" + mappedManagementPort + endpoint)
-                .header("Accept", "application/json")
-                .header("Authorization", credential)
-                .post(RequestBody.create(body.getBytes(StandardCharsets.UTF_8)))
-                .build();
-
-        try {
-            return client.newCall(request).execute();
-        } catch (IOException e) {
-            log.error("Error while making API call to RabbitMQ", e);
-        }
-        return null;
-    }
-
-    private static GenericContainer rabbitMQContainer() {
-        RabbitMQContainerConfiguration containerConfiguration = new RabbitMQContainerConfiguration();
-        containerConfiguration.setDockerImage(RABBITMQ_DOCKER_IMAGE);
-        log.info("Starting rabbitMQ server. Docker image: {}", containerConfiguration.getDockerImage());
-
-        GenericContainer rabbitMQ =
-                new GenericContainer(RABBITMQ_DOCKER_IMAGE)
-                        .withEnv("RABBITMQ_DEFAULT_VHOST", containerConfiguration.getVhost())
-                        .withEnv("RABBITMQ_DEFAULT_USER", RABBITMQ_USERNAME)
-                        .withEnv("RABBITMQ_DEFAULT_PASS", RABBITMQ_PASSWORD)
-                        .withExposedPorts(containerConfiguration.getPort(), RABBITMQ_MANAGEMENT_PORT)
-                        .waitingFor(new RabbitMQStatusCheck(containerConfiguration))
-                        .withStartupTimeout(Duration.ofSeconds(30));
-
-        rabbitMQ = rabbitMQ.withStartupCheckStrategy(new IsRunningStartupCheckStrategyWithDelay());
-        rabbitMQ.start();
-        log.info("Started RabbitMQ server");
-        return rabbitMQ;
     }
 
     private static RMQConfig getRMQConfig(GenericContainer rabbitmqContainer) {